requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backoff==2.2.1
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.4
dataclasses-json==0.5.7
deprecation==2.0.7
distlib==0.3.6
distro==1.9.0
et-xmlfile==1.1.0
filelock==3.12.2
frozenlist==1.3.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx>=0.27.0
httpx-sse==0.4.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-ollama==0.2.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.4
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
ollama==0.4.4
openai==1.58.1
openpyxl==3.1.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.1
pluggy==1.2.0
pre-commit==3.3.3
pybind11==2.11.1
pycairo==1.24.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
PyGObject==3.44.1
pypdf==5.1.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2024.2
PyYAML==6.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rudder-sdk-python==2.0.2
ruff==0.0.272
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
termcolor==2.3.0
tiktoken==0.8.0
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.0.3
virtualenv==20.23.1
yarl==1.9.2
